requirements.txt
bson
marshmallow
mongoengine
marshmallow-mongoengine
hug
PyJWT
pytest
mongomock
gunicorn